pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]

[project]
name = "lees-diffusion-experiments"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy",
    "tqdm",
    "pyyaml",
    "matplotlib",
    "seaborn",
    "torchmetrics",
    "pandas",
    "datasets>=2.13.0"
]